Create BlockchainWatcher base class
This is in perparation for the sentinel service which will have to watch
events and persist the latest indexed state. For this the same logic as
in the validator service can be used.
